+++ ext/opcache/Optimizer/sccp.c
@@ -443,7 +443,7 @@ static inline int ct_eval_fetch_dim(zval *result, zval *op1, zval *op2, int supp
 		if (zval_to_string_offset(&index, op2) == FAILURE) {
 			return FAILURE;
 		}
-		if (index >= 0 && index < Z_STRLEN_P(op1)) {
+		if (index >= 0 && (size_t)index < Z_STRLEN_P(op1)) {
 			ZVAL_STR(result, zend_string_init(&Z_STRVAL_P(op1)[index], 1, 0));
 			return SUCCESS;
 		}
